refactor(model): move DeepSeek V4 cache hooks into model
Remove the DeepSeek V4 compressor worker patch and keep the Ascend cache implementations in the self-maintained DeepSeek V4 model path.
